Use Django's HttpResponseRedirect for redirects
Django's HttpResponseRedirect allows only some URL schemes, like `http`
and `https`. To circumvent this limitation, django-oidc-provider used a
simple HttpResponse class for redirects. But that doesn't play well with
the Django test client which assumes that redirection responses always
contain a `url` attribute [1].

Use Django's HttpResponseRedirect class, but tweak it to accept any
scheme, by discarding the exception it raises when it encounters an
unsupported scheme.

[1] https://github.com/django/django/blob/3.2.17/django/test/client.py#L821